Move localization to its own utility file
